Service2/requirements.txt
pykafka==2.8.0
connexion==2.6.0
Flask_Cors==3.0.8
SQLAlchemy==1.3.13
mysql_connector_repackaged==0.3.1
PyYAML==5.3
pymysql==0.9.3
